+++ src/main/java/net/meander/subtlyd/mixin/client/options/VideoSettingsScreenMixin.java
@@ -1,11 +1,11 @@
 package net.meander.subtlyd.mixin.client.options;
 
-import net.meander.subtlyd.client.OptionInstanceSD;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.meander.subtlyd.client.OptionInstanceSD;
 import net.minecraft.client.OptionInstance;
 import net.minecraft.client.Options;
-import net.minecraft.client.gui.screens.options.AccessibilityOptionsScreen;
+import net.minecraft.client.gui.screens.options.VideoSettingsScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -15,15 +15,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Environment(EnvType.CLIENT)
-@Mixin(AccessibilityOptionsScreen.class)
-public class AccessibilityOptionsScreenMixin {
+@Mixin(VideoSettingsScreen.class)
+public class VideoSettingsScreenMixin {
     /**
-     * Adds new accessibility options.
+     * Adds new video options.
      */
-    @Inject(method = "options", at = @At("RETURN"), cancellable = true)
-    private static void options(Options options, CallbackInfoReturnable<OptionInstance<?>[]> cir) {
+    @Inject(method = "displayOptions", at = @At("RETURN"), cancellable = true)
+    private static void displayOptions(Options options, CallbackInfoReturnable<OptionInstance<?>[]> cir) {
         List<OptionInstance<?>> optionInstanceList = new ArrayList<>(List.of(cir.getReturnValue().clone()));
-        optionInstanceList.add(18, new OptionInstanceSD().screenShake());
+        optionInstanceList.add(8, new OptionInstanceSD().ui());
         cir.setReturnValue(optionInstanceList.toArray(new OptionInstance[0]));
     }
 }